# perf(source): benchmarks on EndpointTargetsFromServices