## What does it do ?

In this PR, I've added a benchmark tests, to showcase bottlenecks described here https://github.com/kubernetes-sigs/external-dns/pull/5473#issuecomment-2911576025

- benchmark tests
- added more tests
- found duplicate code and replaced with function

## Motivation

When decided to have a look started

- https://github.com/kubernetes-sigs/external-dns/pull/5473
- https://github.com/kubernetes-sigs/external-dns/issues/5016

And this comment https://github.com/kubernetes-sigs/external-dns/pull/5493#discussion_r2143457453

One of the bottlenecks is in this lines https://github.com/kubernetes-sigs/external-dns/blob/051c900d1c6073a2e02c2b55c05ab932af8bcc66/source/endpoints.go#L87-L110

Basically it does
- retrive all services from the cache
- iterate over each service
- iterate over each label

On top of that, the function EndpointTargetsFromServices actually called from within the loop of all virtual services or any other kubernetes entity like [here](https://github.com/kubernetes-sigs/external-dns/blob/051c900d1c6073a2e02c2b55c05ab932af8bcc66/source/istio_gateway.go#L143). This is quite inneficiate....

Benchmark results

```
BenchmarkEndpointTargetsFromServicesMedium-16                       	   11194	    105419 ns/op
BenchmarkEndpointTargetsFromServicesMediumIterateOverGateways
BenchmarkEndpointTargetsFromServicesMediumIterateOverGateways-16    	     290	   4070484 ns/op
BenchmarkEndpointTargetsFromServicesHigh
BenchmarkEndpointTargetsFromServicesHigh-16                         	      88	  12669344 ns/op
BenchmarkEndpointTargetsFromServicesHighIterateOverGateways
BenchmarkEndpointTargetsFromServicesHighIterateOverGateways-16      	       1	13425263466 ns/op
```

Just a hightlight from one of the issues
> We are running on EKS with BatchSizeConfig of 1,000 (default), we have ~3K Isito virtualService objects & 1K Service objects.

So this means that there are 3000 * 1000 * (number of selectors), every time there is a reconcile. 

Follo-up:
- potential improvement from this PR https://github.com/gofogo/k8s-sigs-external-dns-fork/pull/31

## More

- [X] Yes, this PR title follows [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/)
- [X] Yes, I added unit tests
- [ ] Yes, I updated end user documentation accordingly

<!--
    Please read https://github.com/kubernetes-sigs/external-dns#contributing before submitting
    your pull request. Please fill in each section below to help us better prioritize your pull request. Thanks!
-->